wii: press key mapping configuration adjust
Currently, the default wiimote key mapping configure 'A' as attack,
'1' as jump.

This makes wii users need to reconfigure the cfg file to make the
key mapping to be more friendly with '1' as attack, '2' as jump.

Change the default key mapping to '1' as attach, '2' as jump for
most of the wiimote users.